pyproject.toml
[tool.poetry]
name = "python-sql"
version = "0.1.0"
description = ""
authors = ["akasmall <aka_sm@ro.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
pytest = "^8.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
